requirements.txt
fastapi
uvicorn[standard]
databases
asyncpg
bcrypt
pydantic
python-jose
pydantic[email]
